rename: slice 폴더명 수정
